split out asset html generation
